Refactored backend modules to mimic     subprocess & io builtin modules:     - ffmpegprocess     - io     - _utils/threaded_pipe     - _utils/queue_pausable     - _utils/pipe_nonblock

+++ src/ffmpegio/streams/SimpleStreams.py
@@ -1,9 +1,17 @@
 import numpy as np
-from .. import ffmpeg, utils, configure
+from .. import ffmpeg, utils, configure, ffmpegprocess
 
 
 class SimpleVideoReader:
-    def __init__(self, url, stream_id=0, **options) -> None:
+    def __init__(
+        self,
+        url,
+        stream_id=0,
+        progress=None,
+        capture_log=None,
+        queue_size=None,
+        **options,
+    ) -> None:
         self.url = url
         self.stream_id = stream_id
 
@@ -27,17 +35,14 @@ class SimpleVideoReader:
             * self.frame_rate
         )
 
-        self._proc = ffmpeg.run(args)
-        self._stdout = self._proc.stdout
-        self._nbytes = (
-            self.shape[0]
-            * self.shape[1]
-            * self.shape[2]
-            * np.dtype(self.dtype).itemsize
+        self._proc = ffmpegprocess.Popen(
+            args,
+            progress=progress,
+            capture_log=capture_log,
+            input_queue_size=queue_size or 16,
         )
-        self.eof = False
 
-    def read(self, vframes=1):
+    def read(self, vframes=1, block=True, timeout=None):
         """read video frames
 
         :param url: audio/video url
@@ -48,33 +53,73 @@ class SimpleVideoReader:
         :return: frame rate and video frame data (dims: time x rows x cols x pix_comps)
         :rtype: (fractions.Fraction, numpy.ndarray)
         """
-        if self.eof:
-            return None
 
-        nbytes = self._nbytes * vframes
-        data = self._stdout.read(nbytes if nbytes > 0 else -1)
-        self.eof = nbytes < 1 or len(data) < nbytes
-        data = np.frombuffer(data, dtype=self.dtype).reshape((-1, *self.shape))
-        self.frame_index += data.shape[0]
-        return data
+        return self._proc.stdout.read_as_array(
+            vframes, block, timeout, self.shape, self.dtype
+        )
 
     def close(self):
-        if not self.eof:
+        if self._proc:
             self._proc.terminate()
-        self._proc.wait()
+            self._proc = None
 
 
 class SimpleVideoWriter:
-    def __init__(self, url, rate, dtype=None, shape=None, **kwargs) -> None:
-        self._proc = self.shape = None
+    def __init__(
+        self,
+        url,
+        rate,
+        dtype=None,
+        size=None,
+        progress=None,
+        input_copy=True,
+        capture_log=None,
+        queue_size=None,
+        **kwargs,
+    ) -> None:
         self.url = url
         self.frame_rate = rate
-        self.dtype = dtype
-        self.shape = shape
+        self.dtype = dtype  # deduces pix_fmt
+        self.size = size  # (w,h)
         self.options = kwargs
         self.frames_written = 0
+        self.progress = progress
+        self.input_copy = input_copy
+        self.capture_log = capture_log
+        self.queue_size = queue_size
 
-    def write(self, data):
+        if dtype and size:
+            self._open(dtype=dtype, size=size)
+        else:
+            self._proc = None
+
+    def _open(self, data=None, dtype=None, size=None):
+        args = configure.input_timing(
+            {},
+            "-",
+            vstream_id=0,
+            excludes=("start", "end", "duration"),
+            **{"input_frame_rate": self.frame_rate, **self.options},
+        )
+
+        configure.video_io(
+            args,
+            utils.array_to_video_input(
+                self.frame_rate, data, format="rawvideo", dtype=dtype, size=size
+            ),
+            output_url=self.url,
+            **self.options,
+        )
+
+        self._proc = ffmpegprocess.Popen(
+            args,
+            progress=self.progress,
+            input_copy=self.input_copy,
+            capture_log=self.capture_log,
+            input_queue_size=self.queue_size or 16,
+        )
+
+    def write(self, data, block=True, timeout=None):
         """write video frames
 
         :param url: audio/video url
@@ -99,36 +144,26 @@ class SimpleVideoWriter:
                 raise Exception("mismatched frame size.")
 
         else:
-            args = configure.input_timing(
-                {},
-                "-",
-                vstream_id=0,
-                excludes=("start", "end", "duration"),
-                **{"input_frame_rate": self.frame_rate, **self.options},
-            )
-
-            configure.video_io(
-                args,
-                utils.array_to_video_input(self.frame_rate, data, format="rawvideo"),
-                output_url=self.url,
-                **self.options,
-            )
-
-            self._proc = ffmpeg.run(args, stdout=None, stdin=ffmpeg.PIPE)
-            self.dtype = data.dtype
-            self.shape = data.shape[-3:]
-
-        self._proc.stdin.write(data.tobytes())
+            self._open(data)
+        self._proc.stdin.write(data, block, timeout)
         self.frames_written += data.shape[0] if data.ndim == 4 else 1
 
     def close(self):
-        if self._proc is not None:
-            self._proc.stdin.close()
-            self._proc.wait()
+        if self._proc:
+            self._proc.terminate()
+            self._proc = None
 
 
 class SimpleAudioReader:
-    def __init__(self, url, stream_id=0, **options) -> None:
+    def __init__(
+        self,
+        url,
+        stream_id=0,
+        progress=None,
+        capture_log=None,
+        queue_size=None,
+        **options,
+    ) -> None:
         self.url = url
         self.stream_id = stream_id
         args = configure.input_timing({}, url, astream_id=stream_id, **options)
@@ -149,21 +184,22 @@ class SimpleAudioReader:
 
         self.dtype, self.channels, self.sample_rate = reader_cfg[0]
 
-        self._proc = ffmpeg.run(args)
-        self._stdout = self._proc.stdout
-        self._nbytes = self.channels * np.dtype(self.dtype).itemsize
+        self._proc = ffmpegprocess.Popen(
+            args,
+            progress=progress,
+            capture_log=capture_log,
+            output_queue_size=queue_size or 16,
+        )
+
         self.remaining = end
-        self.eof = end <= 0
         self.sample_index = start
 
         # if starting mid-stream, drop earlier samples
-        while start > 8192:
-            self.read(8192)
-            start -= 8192
         if start > 0:
-            self.read(start)
+            nblk = self.channels * np.dtype(self.dtype).itemsize
+            self.remaining -= len(self._proc.stdout.read(start * nblk, True)) // nblk
 
-    def read(self, nsamples=0):
+    def read(self, nsamples=0, block=True, timeout=None):
         """read audio samples
 
         :param url: audio/video url
@@ -175,43 +211,65 @@ class SimpleAudioReader:
         :rtype: (fractions.Fraction, numpy.ndarray)
         """
 
-        if self.eof:
+        if not self.remaining:
             return None
 
-        nbytes = self._nbytes * nsamples
-        data = np.frombuffer(
-            self._stdout.read(nbytes if nbytes > 0 else -1), dtype=self.dtype
-        ).reshape((-1, self.channels))
-        if self.remaining <= data.shape[0]:
-            self.eof = True
-            data = data[: self.remaining, :]
+        if nsamples > self.remaining:
+            nsamples = self.remaining
+
+        data = self._proc.stdout.read_as_array(
+            nsamples, block, timeout, self.channels, self.dtype
+        )
         self.remaining -= data.shape[0]
         self.sample_index += data.shape[0]
+
         return data
 
-    def readiter(self, nsamples):
-        while not self.eof:
-            yield self.read(nsamples)
+    def readiter(self, nsamples, block=True, timeout=None):
+        while self.remaining > 0:
+            data = self.read(nsamples, block, timeout)
+            if not data.shape[0]:
+                break  # no data, not data for nonblocking or timed out
+            yield data
 
     def close(self):
-        self._proc.terminate()
-        self._proc.wait()
+        if self._proc:
+            self._proc.terminate()
+            self._proc = None
 
 
 class SimpleAudioWriter:
-    def __init__(self, url, rate, dtype=None, channels=None, **options) -> None:
+    def __init__(
+        self,
+        url,
+        rate,
+        dtype=None,
+        channels=None,
+        progress=None,
+        input_copy=True,
+        capture_log=None,
+        queue_size=None,
+        **options,
+    ) -> None:
         self._proc = None
         self.url = url
         self.samples_written = 0
+        self.progress = progress
+        self.input_copy = input_copy
+        self.capture_log = capture_log
         if dtype is None or channels is None:
             self.sample_rate = rate
             self.dtype = dtype
             self.channels = channels
             self.options = options
         else:
-            self.open(rate, dtype=dtype, channels=channels, **options)
+            self.open(
+                rate, dtype=dtype, channels=channels, queue_size=queue_size, **options
+            )
 
-    def open(self, rate, data=None, dtype=None, channels=None, **options):
+    def open(
+        self, rate, data=None, dtype=None, channels=None, queue_size=None, **options
+    ):
         if self._proc:
             raise Exception("stream is already open")
 
@@ -235,10 +293,16 @@ class SimpleAudioWriter:
             output_url=self.url,
             **self.options,
         )
-        self._proc = ffmpeg.run(args, stdout=None, stdin=ffmpeg.PIPE)
+        self._proc = ffmpegprocess.Popen(
+            args,
+            progress=self.progress,
+            input_copy=self.input_copy,
+            capture_log=self.capture_log,
+            input_queue_size=queue_size or 16,
+        )
         self.samples_written = 0
 
-    def write(self, data):
+    def write(self, data, block=True, timeout=None):
         """write audio data
 
         :param url: audio/video url
@@ -267,10 +331,10 @@ class SimpleAudioWriter:
                 raise Exception("audio data must be 1d or 2d numpy.ndarray")
             self.open(self.sample_rate, data)
 
-        self._proc.stdin.write(data.tobytes())
+        self._proc.stdin.write(data, block, timeout)
         self.samples_written += data.shape[0]
 
     def close(self):
-        if self._proc is not None:
-            self._proc.stdin.close()
-            self._proc.wait()
+        if self._proc:
+            self._proc.terminate()
+            self._proc = None
